requirements.txt
PyQt5>=5.15
PyInstaller>=6.0
Pillow>=10.0
pywin32>=306
olefile>=0.46
